Move to lulpeg, update readme and rockspecs (#4)
We decided to switch to `lulpeg` for now.
It is written in lua and appears to be more stable then lpeg,
which happen to crash under tarantool (or luajit). Additional
research on this topic is needed.

Also, unlike lpeg, lulpeg is available in tarantool ecosystem.

We may return to the topic lpeg/lulpeg later, when performance issues
pops up.

Removed graphql/package.lua, since there are no plans to distribute
package through lit right now.