Support custom directives
GraphQL custom directives was introduced in spec #3.13 in October 2021
release [1]. Directives are the preferred way to extend GraphQL with
custom or experimental behavior. This patch adds support of custom
directives, as well as several location adjustments and repeatable
directive option needed for full support of custom directives.

This patch also adds introspection test for schema with custom
directives to validate schema. Introspection of schema is described in
spec #4.2 [2].

1. https://spec.graphql.org/October2021/#sec-Type-System.Directives.Custom-Directives
2. https://spec.graphql.org/October2021/#sec-Schema-Introspection

Based on PR #20 by @no1seman