Propagate defaults to operation callback
Before this PR there is no fast and convenient way to get
operation and directives defaults set when creating the schema.

Now it's possible, operation and directives default can be easily found
is in the #3 argument of callback - map "info". The keys are:

defaultValues - for operation defaults
directivesDefaultValues - for directives defaults